Add module for display hero
